+++ hydromt/model/__init__.py
@@ -1,7 +1,13 @@
 """HydroMT Model API."""
 
+from hydromt.model.example.example_model import ExampleModel
 from hydromt.model.model import Model
 from hydromt.model.root import ModelRoot
 from hydromt.model.steps import hydromt_step
 
-__all__ = ["Model", "hydromt_step", "ModelRoot"]
+__all__ = ["Model", "ExampleModel", "ModelRoot", "hydromt_step"]
+
+
+# define hydromt model entry points
+# see also hydromt.model group in pyproject.toml
+__hydromt_eps__ = ["Model", "ExampleModel"]